split harn-fmt formatter implementation
